Extract base test class
